Move button to core
