+++ packages/mobile/src/components/core/DynamicImage.tsx
