--- packages/mobile/src/components/core/DynamicImage/index.ts
@@ -1,1 +1,0 @@
-export { DynamicImage } from './DynamicImage'
