+++ packages/mobile/src/components/core/index.ts
@@ -1,7 +1,8 @@
-export * from './IconButton'
-export * from './Tile'
-export * from './DynamicImage'
+export * from './Button'
 export * from './CardList'
+export * from './DynamicImage'
 export * from './EmptyCard'
+export * from './IconButton'
 export * from './TextButton'
+export * from './Tile'
 export * from './VirtualizedScrollView'
