+++ packages/mobile/src/screens/profile-screen/SubscribeButton.tsx
@@ -1,8 +1,8 @@
 import IconNotification from 'app/assets/images/iconNotification.svg'
+import { Button } from 'app/components/core'
 import { useSelectorWeb } from 'app/hooks/useSelectorWeb'
 import { makeStyles } from 'app/styles'
 
-import { Button } from './Button'
 import { getProfile } from './selectors'
 
 const messages = {
